+++ docs/changelog.md
@@ -10,6 +10,8 @@ Sprint 10 / April 12, 2020
 
 ### Updates
 
+- added total daily time to Dashboard.
+
 <hr />
 
 ## Release 0.9.0